The whole project is restructured and initiated again
